Исправлена ошибка в формуле и типе данных константы.

+++ src/trainer/consistenttrainer.h
@@ -44,7 +44,7 @@ public:
     // суффикс именования файла с данными о тренировке модели
     static QString TRAINER_DATA_NAME;
     // коэффициент мощности("тяжеловесности") модели
-    static double MODEL_POWER_FACTOR;
+    static long long MODEL_POWER_FACTOR;
 
     bool operator==(const ConsistentTrainer &trainer);
